Move ConfirmedCasesDetailsCard to components/index/cards
