+++ components/index/cards/ConfirmedCasesDetailsCard.vue
